(update) file structure and consolidated code
